requirements.txt
appdirs==1.4.4
Babel==2.11.0
beautifulsoup4==4.11.1
black==22.10.0
bs4==0.0.1
certifi==2022.9.24
click==8.1.3
cssselect==1.1.0
dateparser==1.1.1
demjson3==3.0.5
distlib==0.3.6
facebook-scraper==0.2.59
fake-useragent==0.1.11
filelock==3.8.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-SQLAlchemy==3.0.2
Flask-Table==0.5.0
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
mpmath==1.2.1
mypy-extensions==0.4.3
numpy==1.23.5
parse==1.19.0
pathspec==0.10.2
pip==22.2.2
platformdirs==2.5.2
psycopg2-binary==2.9.5
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
setuptools==65.5.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sympy==1.11.1
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.5
tzlocal==4.2
virtualenv==20.16.5
w3lib==2.0.1
websockets==10.3
Werkzeug==2.2.2
wheel==0.37.1
